requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pandas==2.1.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==1.2.0
